spargebra

Module term

Data structures for RDF 1.1 Concepts like IRI, literal or triples.

Structs

BlankNode

An owned RDF blank node.

GroundQuad

A RDF triple in an RDF dataset without blank nodes.

GroundQuadPattern

A triple pattern in a specific graph without blank nodes.

GroundTriple

A RDF triple without blank nodes.

GroundTriplePattern

A triple pattern without blank nodes.

Literal

An owned RDF literal.

NamedNode

An owned RDF IRI.

Quad

A RDF triple in an RDF dataset.

QuadPattern

A triple pattern in a specific graph

Triple

An owned RDF triple.

TriplePattern

A triple pattern

Variable

A SPARQL query owned variable.

Enums

GraphName

A possible graph name.

GraphNamePattern

The union of IRIs, default graph name and variables.

GroundSubject

The union of IRIs and triples.

GroundTerm

The union of IRIs, literals and triples.

GroundTermPattern

The union of terms and variables without blank nodes.

NamedNodePattern

The union of IRIs and variables.

Subject

The owned union of IRIs, blank nodes and triples (if the rdf-star feature is enabled).

Term

An owned RDF term

TermPattern

The union of terms and variables.