spargebra::term
Struct Literal

pub struct Literal(/* private fields */);

An owned RDF literal.

The default string formatter is returning an N-Triples, Turtle, and SPARQL compatible representation:

use oxrdf::vocab::xsd;
use oxrdf::Literal;

assert_eq!(
    "\"foo\\nbar\"",
    Literal::new_simple_literal("foo\nbar").to_string()
);

assert_eq!(
    r#""1999-01-01"^^<http://www.w3.org/2001/XMLSchema#date>"#,
    Literal::new_typed_literal("1999-01-01", xsd::DATE).to_string()
);

assert_eq!(
    r#""foo"@en"#,
    Literal::new_language_tagged_literal("foo", "en")?.to_string()
);

Implementations

impl Literal

pub fn new_simple_literal(value: impl Into<String>) -> Literal

Builds an RDF simple literal.

pub fn new_typed_literal( value: impl Into<String>, datatype: impl Into<NamedNode>, ) -> Literal

Builds an RDF literal with a datatype.

pub fn new_language_tagged_literal( value: impl Into<String>, language: impl Into<String>, ) -> Result<Literal, LanguageTagParseError>

Builds an RDF language-tagged string.

pub fn new_language_tagged_literal_unchecked( value: impl Into<String>, language: impl Into<String>, ) -> Literal

Builds an RDF language-tagged string.

It is the responsibility of the caller to check that language is valid BCP47 language tag, and is lowercase.

Literal::new_language_tagged_literal() is a safe version of this constructor and should be used for untrusted data.

pub fn value(&self) -> &str

The literal lexical form.

pub fn language(&self) -> Option<&str>

The literal language tag if it is a language-tagged string.

Language tags are defined by the BCP47. They are normalized to lowercase by this implementation.

pub fn datatype(&self) -> NamedNodeRef<'_>

The literal datatype.

The datatype of language-tagged string is always rdf:langString. The datatype of simple literals is xsd:string.

pub fn is_plain(&self) -> bool

Checks if this literal could be seen as an RDF 1.0 plain literal.

It returns true if the literal is a language-tagged string or has the datatype xsd:string.

pub fn as_ref(&self) -> LiteralRef<'_>

pub fn destruct(self) -> (String, Option<NamedNode>, Option<String>)

Extract components from this literal (value, datatype and language tag).